update the logic of delete in UserServiceImpl

+++ RestApi/src/main/java/in/sp/RestApi/service/UserServiceImpl.java
@@ -52,6 +52,9 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(int id) {
+        if(!userRepository.existsById(id)){
+            throw new UserNotFoundException(id+" is not found");
+        }
         userRepository.deleteById(id);
     }
 }
